third_party/blink/web_tests/platform/generic/fast/forms/checkValidity-select-form-expected.txt

This test checks if checkValidity() returns correctly a true result on selecting a valid element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS form.checkValidity() is true
PASS successfullyParsed is true

TEST COMPLETE

abc
